lookup_address: Always return an Address
Instead of returning `None` when `lookup_address` cannot find an address
in the process address space, return an 'invalid' `Address` object. This
forces the callers check for the `valid` attribute instead of just
`None`, but also allows them to check the original address (`value`).

This fixes a bug in `pcustom` where if a struct's field looked like a
pointer but wasn't it showed up as zero.