Issue #90: Fixing ida-interact sync() problem when receiving gdb.ContinueEvent.
There is a race-condition in GDB when a gdb.ContinueEvent is emitted: when the
signal is sent GDB changes the current thread mode to `running` immediately. If
the signal handler receives the signal after the mode is set back to running, it
will prevent to get register values, or set new BP (as done by synchronize()).

We use a small hack to force an additional delay before continuing.