gdb_disassemble: Split on any whitespace, not just literal space
When disassembling AARCH64 gdb puts a \t between the mnemonic and the
first operand. If there is more than operand that is no issue, since we
detect the space after the first and split the whole string. If the
instruction, on the other hand, has only one operand, then we do not
detect it and do not split.

This change simply splits the instruction, and if the split does
anything, there must be operands.